4 may

+++ candidates/models.py
@@ -91,16 +91,19 @@ Apply Button
 class Candidate(models.Model):
     Name = models.CharField(max_length=100,)
     Email_id = models.CharField(max_length=100,null = True,blank=True)
-    Years_of_Experience = models.IntegerField(null = True, blank=True, default=None)
+    Years_of_Experience = models.DecimalField(null = True, blank=True, default=None,max_digits=1,decimal_places=0 )
     Linkedin_Profile = models.CharField(max_length=100,null = True)
-    Expected_hourly_rate = models.IntegerField(null = True,blank=True,default=None)
+    Expected_hourly_rate = models.DecimalField(null = True, blank=True, default=None,max_digits=2,decimal_places=0 )
     Resume = models.FileField(upload_to='doc', validators=[file_size])
     job=models.ForeignKey(Role, on_delete=models.CASCADE,null=True)
     req=models.ForeignKey(Job, on_delete=models.CASCADE,null=True)
+
     def __str__(self):
         return self.Name
+
     def Req_Id(self):
         return self.req.Req_Id
+
 
 class CandidateJobMap(models.Model):
     candidate=models.ForeignKey(Candidate, on_delete=models.CASCADE)
@@ -114,6 +117,7 @@ class CandidateJobMap(models.Model):
 
     def Years_of_Experience(self):
         return self.candidate.Years_of_Experience
+
     def Email_id(self):
         return self.candidate.Email_id
